chore: rename order/controller package to api/order
