+++ src/main/java/com/example/minimart/api/order/dto/request/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package com.example.minimart.order.controller.dto.request;
+package com.example.minimart.api.order.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
